requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
Faker==19.2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==21.2.0
django-grappelli==3.0.6